Share logic in mailers
* Define `subject` in ApplicationMailer
* Use new layout as deafult layout
