+++ app/mailers/attachment_mailer.rb
@@ -1,18 +1,15 @@
 class AttachmentMailer < ApplicationMailer
-  layout "basic"
-
   def new_attachment_notification(to_email, proposal, attachment)
     @proposal = proposal.decorate
     @attachment_user = attachment.user
     @attachment = attachment
-    assign_threading_headers(proposal)
-    subject = "Request #{proposal.public_id}: #{proposal.name}"
+    assign_threading_headers(@proposal)
 
     mail(
       to: to_email,
-      subject: subject,
+      subject: subject(@proposal),
       from: user_email_with_name(@attachment_user),
-      reply_to: reply_email(proposal)
+      reply_to: reply_email(@proposal)
     )
   end
 end
