+++ app/mailers/observer_mailer.rb
@@ -1,33 +1,28 @@
 class ObserverMailer < ApplicationMailer
-  layout "mailer"
-  add_template_helper ValueHelper
-
   def on_observer_added(observation, reason)
     @observation = observation
     @reason = reason
-    observer = observation.user
+    observer = @observation.user
+    @proposal =  observation.proposal.decorate
+    assign_threading_headers(@proposal)
 
-    send_proposal_email(
-      from_email: observation_added_from(observation),
-      to_email: observer.email_address,
-      proposal: observation.proposal
+    mail(
+      to: observer.email_address,
+      from: default_sender_email,
+      subject: subject(@proposal),
+      reply_to: reply_email(@proposal)
     )
   end
 
   def proposal_observer_email(to_email, proposal)
-    send_proposal_email(
-      to_email: to_email,
-      proposal: proposal
+    @proposal = proposal.decorate
+    assign_threading_headers(@proposal)
+
+    mail(
+      to: to_email,
+      from: default_sender_email,
+      subject: subject(@proposal),
+      reply_to: reply_email(@proposal)
     )
-  end
-
-  private
-
-  def observation_added_from(observation)
-    adder = observation.created_by
-
-    if adder
-      user_email_with_name(adder)
-    end
   end
 end
