+++ app/mailers/welcome_mailer.rb
@@ -1,6 +1,4 @@
 class WelcomeMailer < ApplicationMailer
-  layout "basic"
-
   def welcome_notification(user)
     mail(
       to: email_with_name(user.email_address, user.full_name),
