+++ app/views/layouts/email.html.erb
