+++ app/views/step_mailer/step_reply_received.html.haml
@@ -1,21 +1,16 @@
 %table.reply-section
-  - if @proposal.approved?
-    %tr.success
-      %td{ colspan: "2" }
-        = complete_text(@step)
-  - else
-    %tr
-      %td.w-container.html-email-message
-        %p
-          = t("mailer.generic_greeting")
-          ,
-          %br
-          = t("mailer.step_mailer.approval_reply_received.body",
-            full_name: @step.user.full_name,
-            status: @step.status,
-            public_id: @proposal.public_id)
-          %br
-          = t("mailer.step_mailer.approval_reply_received.details")
+  %tr
+    %td.w-container.html-email-message
+      %p
+        = t("mailer.generic_greeting")
+        ,
+        %br
+        = t("mailer.step_mailer.step_reply_received.body",
+          full_name: @step.user.full_name,
+          status: @step.status,
+          public_id: @proposal.public_id)
+        %br
+        = t("mailer.step_mailer.step_reply_received.details")
 
-      = render(partial: "mail_shared/email_reply",
-        locals: { step: @step, show_step_actions: false, proposal: @proposal })
+    = render(partial: "mail_shared/email_reply",
+      locals: { step: @step, show_step_actions: false, proposal: @proposal })
